DS-3484: Moved DSpaceCurieProvider to HATEOAS package
